requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.6
django-debug-toolbar==3.2.2
django-model-utils==4.1.1
djangorestframework==3.12.4
drf-spectacular==0.18.1
inflection==0.5.1
jsonschema==3.2.0
kombu==5.1.0
names==0.3.0
prompt-toolkit==3.0.20
pyrsistent==0.18.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
vine==5.0.0
wcwidth==0.2.5
